Backend/ai_vis_detect/requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
django-cors-headers==4.2.0
django-rest-framework==0.1.0
Django==4.2.4
djangorestframework==3.14.0
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
google-auth-oauthlib==0.4.6
google-auth==2.22.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
joblib==1.3.2
Keras-Preprocessing==1.1.2
keras==2.8.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.0
pip==21.3.1
protobuf==3.19.6
pyasn1-modules==0.3.0
pyasn1==0.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests-oauthlib==1.3.1
requests==2.31.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
setuptools==60.2.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.4
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.8.0
tensorflow-estimator==2.8.0
tensorflow-hub==0.14.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-text==2.8.2
tensorflow==2.8.4
termcolor==2.3.0
threadpoolctl==3.2.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
werkzeug==2.3.7
wheel==0.37.1
wrapt==1.15.0
zipp==3.16.2
